+++ cli/internal/terraform/terraform/gcp/outputs.tf
@@ -1,0 +1,3 @@
+output "ip" {
+  value = google_compute_global_address.loadbalancer_ip.address
+}